feat: implement rbac

+++ RR.App/Controllers/HRIS/EmployeeRoleManageController.cs
@@ -13,13 +13,15 @@ public class EmployeeRoleManageController : ControllerBase
     private readonly IEmployeeRoleService? _employeeRoleService;
     private readonly IEmployeeService? _employeeService;
     private readonly IRoleService? _roleService;
+    private readonly IAuthService _authService;
 
     public EmployeeRoleManageController(IEmployeeRoleService? employeeRoleService, IEmployeeService? employeeService,
-                                        IRoleService? roleService)
+                                        IRoleService? roleService, IAuthService authService)
     {
         _employeeRoleService = employeeRoleService;
         _employeeService = employeeService;
         _roleService = roleService;
+        _authService = authService;
     }
 
     [Authorize(Policy = "AdminOrTalentOrJourneyOrSuperAdminPolicy")]
@@ -28,7 +30,7 @@ public class EmployeeRoleManageController : ControllerBase
     {
         if (string.IsNullOrEmpty(email) || string.IsNullOrEmpty(role)) return BadRequest("Invalid input");
 
-        if(_employeeRoleService == null || _employeeService == null || _roleService == null)
+        if(_employeeRoleService == null || _employeeService == null || _roleService == null || _authService == null)
         {
             return BadRequest("Invalid input");
         }
@@ -36,11 +38,22 @@ public class EmployeeRoleManageController : ControllerBase
         try
         {
             var employee = await _employeeService.GetEmployee(email);
+            if (employee == null)
+            {
+                return NotFound("Employee not found.");
+            }
+
+            var authUser = await _authService.GetUsersByEmailAsync(email);
+            if (authUser == null || !authUser.Any())
+            {
+                return NotFound("User not found in authentication service.");
+            }
+            var authEmployeeId = authUser.First().UserId;
 
             var currRole = await _roleService.CheckRole(role)
                 ? await _roleService.GetRole(role)
                 : await _roleService.SaveRole(new RoleDto { Id = 0, Description = role });
-
+            await _authService.AddRoleToUserAsync(authEmployeeId, currRole.AuthRoleId);
             var employeeRole = new EmployeeRoleDto{ Id = 0, Employee = employee, Role = currRole };
 
             var employeeRoleSaved = await _employeeRoleService.SaveEmployeeRole(employeeRole);
@@ -57,7 +70,7 @@ public class EmployeeRoleManageController : ControllerBase
     [HttpPut]
     public async Task<IActionResult> UpdateRole([FromQuery] string email, [FromQuery] string role)
     {
-        if (_employeeRoleService == null || _employeeService == null || _roleService == null)
+        if (_employeeRoleService == null || _employeeService == null || _roleService == null || _authService == null)
         {
             return BadRequest("Invalid input");
         }
@@ -65,11 +78,30 @@ public class EmployeeRoleManageController : ControllerBase
         try
         {
             var employee = await _employeeService.GetEmployee(email);
+            if (employee == null)
+            {
+                return NotFound("Employee not found.");
+            }
+
+            var authUser = await _authService.GetUsersByEmailAsync(email);
+            if (authUser == null || !authUser.Any())
+            {
+                return NotFound("User not found in authentication service.");
+            }
+
+            var authEmployeeId = authUser.First().UserId;
+
+            var authRoles = await _authService.GetUserRolesAsync(authEmployeeId);
+            if (authRoles != null && authRoles.Any())
+            {
+                var authCurrRoleId = authRoles.First().Id;
+                await _authService.RemoveRoleFromUserAsync(authEmployeeId, authCurrRoleId);
+            }
 
             var currRole = await _roleService.CheckRole(role)
                 ? await _roleService.GetRole(role)
                 : await _roleService.SaveRole(new RoleDto{ Id = 0, Description = role });
-
+            await _authService.AddRoleToUserAsync(authEmployeeId, currRole.AuthRoleId);
             var currEmployeeRole = await _employeeRoleService.GetEmployeeRole(email);
             var employeeRole = new EmployeeRoleDto { Id = currEmployeeRole.Id, Employee = employee, Role = currRole };
 
@@ -89,7 +121,7 @@ public class EmployeeRoleManageController : ControllerBase
     [HttpDelete]
     public async Task<IActionResult> RemoveRole([FromQuery] string email, [FromQuery] string role)
     {
-        if (_employeeRoleService == null || _employeeService == null || _roleService == null)
+        if (_employeeRoleService == null || _employeeService == null || _roleService == null || _authService == null)
         {
             return BadRequest("Invalid input");
         }
@@ -97,10 +129,35 @@ public class EmployeeRoleManageController : ControllerBase
         try
         {
             var employee = await _employeeService.GetEmployee(email);
+            if (employee == null)
+            {
+                return NotFound("Employee not found.");
+            }
 
             var roleToRemove = await _roleService.GetRole(role);
+            if (roleToRemove == null)
+            {
+                return NotFound("Role not found.");
+            }
+
+            var authUser = await _authService.GetUsersByEmailAsync(email);
+            if (authUser == null || !authUser.Any())
+            {
+                return NotFound("User not found in authentication service.");
+            }
+            var authEmployeeId = authUser.First().UserId;
+            var authRoles = await _authService.GetUserRolesAsync(authEmployeeId);
+            var authRoleToRemove = authRoles.FirstOrDefault(r => r.Id == roleToRemove.AuthRoleId);
+            if (authRoleToRemove != null)
+            {
+                await _authService.RemoveRoleFromUserAsync(authEmployeeId, authRoleToRemove.Id);
+            }
 
             var employeeRole = await _employeeRoleService.GetEmployeeRole(email);
+            if (employeeRole == null)
+            {
+                return NotFound("Employee role not found.");
+            }
 
             var employeeRoleRemoved = await _employeeRoleService.DeleteEmployeeRole(email, role);
